Rewrite extra mappings parsing and merging
Instead of relying on a hack to match methods regardless of signature,
this implementation now tries to resolve method signatures upfront,
warning when they can not be resolved or are ambiguous.

This does have the disadvantage that they will no longer function when a
signature cannot be determined (such as e.g. for third-party classes
that are not mapped at all), however the new merging implementation is
well defined and fixes a whole bunch of edge cases that the old one had,
so I'd say it's worth it.